track_playlist retour message si erreur

+++ src/main/java/co/simplon/jukebox/controller/TrackController.java
@@ -3,7 +3,10 @@ package co.simplon.jukebox.controller;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,16 +18,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import co.simplon.jukebox.model.Track;
 import co.simplon.jukebox.service.TrackService;
 
-import javax.validation.Valid;
-
 
 @RestController
 @RequestMapping("/jukebox")
-public class TrackController {
+public class TrackController extends AppController{
 
 	@Autowired
 	TrackService service;
@@ -89,9 +91,11 @@ public class TrackController {
 		if(track.isEmpty())
 			return ResponseEntity.notFound().build();
 		
+		if (!track.get().getPlaylists().isEmpty())
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Track in playlist");
+		
 		service.delete(track.get().getId());
 		return ResponseEntity.accepted().build();
 	}
-
-
+	
 }
